refactor: following the start lane is now possible
The car now follows the starter lane (middle lane). This is achieved by
simply incrementing the Frenet s coordinate and keeping the d coordinate
constant. The trajectory is not smooth and the increment too high,
resulting in very high jerk and acceleration values. This needs to be
fixed in a future commit.